feat: add random city to LoginScreen
