+++ project/src/components/form-login/form-login.tsx
@@ -1,0 +1,70 @@
+import {FormEvent, useRef} from 'react';
+import {useAppDispatch} from '../../hooks';
+import {useNavigate} from 'react-router-dom';
+import {AuthDataType} from '../../types/user';
+import {loginAction} from '../../store/api-actions';
+import {AppRoute} from '../../constants';
+
+const FormLogin = (): JSX.Element => {
+  const loginRef = useRef<HTMLInputElement | null>(null);
+  const passwordRef = useRef<HTMLInputElement | null>(null);
+
+  const dispatch = useAppDispatch();
+  const navigate = useNavigate();
+
+  const onSubmit = (authData: AuthDataType) => {
+    dispatch(loginAction(authData));
+    navigate(AppRoute.Root);
+  };
+
+  const handleSubmit = (evt: FormEvent<HTMLFormElement>) => {
+    evt.preventDefault();
+
+    if (loginRef.current !== null && passwordRef.current !== null) {
+      onSubmit({
+        login: loginRef.current.value,
+        password: passwordRef.current.value,
+      });
+    }
+  };
+
+  return (
+    <form
+      className="login__form form"
+      action=""
+      method=""
+      onSubmit={handleSubmit}
+    >
+      <div className="login__input-wrapper form__input-wrapper">
+        <label className="visually-hidden">E-mail</label>
+        <input
+          className="login__input form__input"
+          type="email"
+          name="email"
+          placeholder="Email"
+          required
+          ref={loginRef}
+        />
+      </div>
+      <div className="login__input-wrapper form__input-wrapper">
+        <label className="visually-hidden">Password</label>
+        <input
+          className="login__input form__input"
+          type="password"
+          name="password"
+          placeholder="Password"
+          required
+          ref={passwordRef}
+        />
+      </div>
+      <button
+        className="login__submit form__submit button"
+        type="submit"
+      >
+        Sign in
+      </button>
+    </form>
+  );
+};
+
+export default FormLogin;
