+++ project/src/components/tab-item/tab-item.tsx
@@ -15,7 +15,7 @@ const TabItem = ({city}: TabItemProps): JSX.Element => {
   });
 
   return (
-    <li className="locations__item" onClick={() => dispatch(changeActiveCity({city: city}))}>
+    <li className="locations__item" onClick={() => dispatch(changeActiveCity(city))}>
       <a className={linkClassName} href="#">
         <span>{city}</span>
       </a>
