+++ project/src/pages/login-screen/login-screen.tsx
@@ -1,33 +1,14 @@
-import {useRef, FormEvent} from 'react';
-import {generatePath, Link, useNavigate} from 'react-router-dom';
+import {generatePath, Link} from 'react-router-dom';
 import {AppRoute} from '../../constants';
-import {useAppDispatch} from '../../hooks';
-import {loginAction} from '../../store/api-actions';
-import {AuthDataType} from '../../types/user';
 import HeaderLogin from '../../components/header-login/header-login';
+import FormLogin from '../../components/form-login/form-login';
+import {CITIES} from '../../constants';
+import {changeActiveCity} from '../../store/action';
+import {useAppDispatch} from '../../hooks';
 
 const LoginScreen = () => {
-  const loginRef = useRef<HTMLInputElement | null>(null);
-  const passwordRef = useRef<HTMLInputElement | null>(null);
-
+  const randomCity = CITIES[Math.floor(Math.random() * CITIES.length)];
   const dispatch = useAppDispatch();
-  const navigate = useNavigate();
-
-  const onSubmit = (authData: AuthDataType) => {
-    dispatch(loginAction(authData));
-    navigate(AppRoute.Root);
-  };
-
-  const handleSubmit = (evt: FormEvent<HTMLFormElement>) => {
-    evt.preventDefault();
-
-    if (loginRef.current !== null && passwordRef.current !== null) {
-      onSubmit({
-        login: loginRef.current.value,
-        password: passwordRef.current.value,
-      });
-    }
-  };
 
   return (
     <>
@@ -36,46 +17,16 @@ const LoginScreen = () => {
         <div className="page__login-container container">
           <section className="login">
             <h1 className="login__title">Sign in</h1>
-            <form
-              className="login__form form"
-              action=""
-              method=""
-              onSubmit={handleSubmit}
-            >
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">E-mail</label>
-                <input
-                  className="login__input form__input"
-                  type="email"
-                  name="email"
-                  placeholder="Email"
-                  required
-                  ref={loginRef}
-                />
-              </div>
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">Password</label>
-                <input
-                  className="login__input form__input"
-                  type="password"
-                  name="password"
-                  placeholder="Password"
-                  required
-                  ref={passwordRef}
-                />
-              </div>
-              <button
-                className="login__submit form__submit button"
-                type="submit"
-              >
-                Sign in
-              </button>
-            </form>
+            <FormLogin/>
           </section>
           <section className="locations locations--login locations--current">
             <div className="locations__item">
-              <Link className="locations__item-link" to={generatePath(AppRoute.Root)}>
-                <span>Amsterdam</span>
+              <Link
+                className="locations__item-link"
+                to={generatePath(AppRoute.Root)}
+                onClick={() => dispatch(changeActiveCity(randomCity))}
+              >
+                <span>{randomCity}</span>
               </Link>
             </div>
           </section>
